Testing Duplicates
I second guessed if 'Record_Number' should be included in duplicate checks, so I ran the code to remove duplicates with that included. I then counted the number of records in the new file, which was the same as the original one. This suggested Record_Number should be removed for duplicate analysis.